upgraded to latest versions
A security issue has been found in the net/http library of the Go language that affects all versions and all components of Kubernetes. The vulnerabilities can result in a DoS against any process with an HTTP or HTTPS listener. 
 
Am I vulnerable?
Yes. All versions of Kubernetes are affected.
 
Go has released versions go1.12.8 and go1.11.13, and we have released the following versions of Kubernetes built using patched versions of Go.

 ·         Kubernetes v1.15.3 - go1.12.9

·         Kubernetes v1.14.6 - go1.12.9

·         Kubernetes v1.13.10 - go1.11.13

https://groups.google.com/forum/m/#!topic/kubernetes-announce/p-c33PN6pzw